third_party/blink/web_tests/webexposed/intl-duration-format-expected.txt

This test dumps the output of Intl.DurationFormat.prototype.format() for
some common locales.

[en]
  [digital] 2:00:35
  [narrow] 2h 35s
  [short] 2 hr, 35 sec
  [long] 2 hours, 35 seconds
[en-GB]
  [digital] 2:00:35
  [narrow] 2h 35s
  [short] 2 hrs, 35 secs
  [long] 2 hours, 35 seconds
[en-US]
  [digital] 2:00:35
  [narrow] 2h 35s
  [short] 2 hr, 35 sec
  [long] 2 hours, 35 seconds
[zh-CN]
  [digital] 2:00:35
  [narrow] 2小时35秒
  [short] 2小时35秒
  [long] 2小时35秒钟
[zh-TW]
  [digital] 2:00:35
  [narrow] 2 小時35 秒
  [short] 2 小時 35 秒
  [long] 2 小時 35 秒
[hi]
  [digital] 2:00:35
  [narrow] 2 घं 35 से
  [short] 2 घं॰, 35 से॰
  [long] 2 घंटे और 35 सेकंड
[es]
  [digital] 2:00:35
  [narrow] 2h 35s
  [short] 2 h y 35 s
  [long] 2 horas y 35 segundos
[es-ES]
  [digital] 2:00:35
  [narrow] 2h 35s
  [short] 2 h y 35 s
  [long] 2 horas y 35 segundos
[es-MX]
  [digital] 2:00:35
  [narrow] 2h 35s
  [short] 2 h y 35 s
  [long] 2 horas y 35 segundos
[es-419]
  [digital] 2:00:35
  [narrow] 2h 35s
  [short] 2 h y 35 s
  [long] 2 horas y 35 segundos
[ar]
  [digital] 2:00:35
  [narrow] 2 س و35 ث
  [short] 2 س و35 ث
  [long] ساعتان و35 ثانية
[fr]
  [digital] 2:00:35
  [narrow] 2h 35s
  [short] 2 h et 35 s
  [long] 2 heures et 35 secondes
[bn]
  [digital] ২:০০:৩৫
  [narrow] ২ ঘঃ, ৩৫ সেঃ
  [short] ২ ঘন্টা, ৩৫ সেকেন্ড
  [long] ২ ঘন্টা, ৩৫ সেকেন্ড
[pt-PT]
  [digital] 2:00:35
  [narrow] 2 h e 35 s
  [short] 2 h e 35 s
  [long] 2 horas e 35 segundos
[pt-BR]
  [digital] 2:00:35
  [narrow] 2 h 35 s
  [short] 2 h e 35 s
  [long] 2 horas e 35 segundos
[ru]
  [digital] 2:00:35
  [narrow] 2 ч 35 с
  [short] 2 ч 35 с
  [long] 2 часа 35 секунд
[id]
  [digital] 2.00.35
  [narrow] 2 j, 35 dtk
  [short] 2 j, 35 dtk
  [long] 2 jam, 35 detik
[ur]
  [digital] 2:00:35
  [narrow] 2 گھنٹے اور 35 سیکنڈ
  [short] 2 گھنٹے اور 35 سیکنڈ
  [long] 2 گھنٹے، 35 سیکنڈ
[de]
  [digital] 2:00:35
  [narrow] 2 Std., 35 Sek.
  [short] 2 Std., 35 Sek.
  [long] 2 Stunden, 35 Sekunden
[ja]
  [digital] 2:00:35
  [narrow] 2h35s
  [short] 2 時間 35 秒
  [long] 2 時間 35 秒
[ar-EG]
  [digital] ٢:٠٠:٣٥
  [narrow] ٢ س و٣٥ ث
  [short] ٢ س و٣٥ ث
  [long] ساعتان و٣٥ ثانية
[mr]
  [digital] २:००:३५
  [narrow] २ता ३५से
  [short] २ ता, ३५ से
  [long] २ तास, ३५ सेकंद
[vi]
  [digital] 2:00:35
  [narrow] 2 giờ 35 giây
  [short] 2 giờ, 35 giây
  [long] 2 giờ, 35 giây
[te]
  [digital] 2:00:35
  [narrow] 2గం, 35సె
  [short] 2 గం., 35 సెక.
  [long] 2 గంటలు, 35 సెకన్లు
[tr]
  [digital] 2:00:35
  [narrow] 2s 35sn
  [short] 2 sa. 35 sn.
  [long] 2 saat 35 saniye
[sw]
  [digital] 2:00:35
  [narrow] saa 2 na sek 35
  [short] saa 2 na sekunde 35
  [long] saa 2 na sekunde 35
[tl]
  [digital] 2:00:35
  [narrow] 2 oras 35s
  [short] 2 oras, 35 seg.
  [long] 2 oras, 35 segundo
[fil]
  [digital] 2:00:35
  [narrow] 2 oras 35s
  [short] 2 oras, 35 seg.
  [long] 2 oras, 35 segundo
[ta]
  [digital] 2:00:35
  [narrow] 2 ம.நே. 35 வி.
  [short] 2 மணிநேரம், 35 விநாடி
  [long] 2 மணிநேரங்கள், 35 விநாடிகள்
[zh-HK]
  [digital] 2:00:35
  [narrow] 2小時35秒
  [short] 2 小時 35 秒
  [long] 2 小時 35 秒
[fa]
  [digital] ۲:۰۰:۳۵
  [narrow] ۲h ۳۵s
  [short] ۲ ساعت،‏ ۳۵ ثانیه
  [long] ۲ ساعت و ۳۵ ثانیه
[ko]
  [digital] 2:00:35
  [narrow] 2시간 35초
  [short] 2시간 35초
  [long] 2시간 35초
[th]
  [digital] 2:00:35
  [narrow] 2ชม. 35วิ
  [short] 2 ชม. 35 วิ
  [long] 2 ชั่วโมง และ 35 วินาที
[it]
  [digital] 2:00:35
  [narrow] 2h 35s
  [short] 2 h e 35 s
  [long] 2 ore e 35 secondi
[gu]
  [digital] 2:00:35
  [narrow] 2 ક, 35 સે
  [short] 2 કલાક, 35 સેકંડ
  [long] 2 કલાક અને 35 સેકંડ
[ar-SY]
  [digital] ٢:٠٠:٣٥
  [narrow] ٢ س و٣٥ ث
  [short] ٢ س و٣٥ ث
  [long] ساعتان و٣٥ ثانية
[am]
  [digital] 2:00:35
  [narrow] 2 ሰ 35 ሰ
  [short] 2 ሰዓ፣ 35 ሰከ
  [long] 2 ሰዓቶች፣ 35 ሰከንዶች
[kn]
  [digital] 2:00:35
  [narrow] 2ಗಂ. 35 ಸೆಕೆಂ
  [short] 2 ಗಂ., 35 ಸೆಕೆಂ
  [long] 2 ಗಂಟೆಗಳು, 35 ಸೆಕೆಂಡುಗಳು
[ar-SD]
  [digital] ٢:٠٠:٣٥
  [narrow] ٢ س و٣٥ ث
  [short] ٢ س و٣٥ ث
  [long] ساعتان و٣٥ ثانية
[nl]
  [digital] 2:00:35
  [narrow] 2 u, 35 s
  [short] 2 uur, 35 sec
  [long] 2 uur en 35 seconden
[pl]
  [digital] 2:00:35
  [narrow] 2 h i 35 s
  [short] 2 godz. i 35 sek.
  [long] 2 godziny i 35 sekund
[sv-SE]
  [digital] 2:00:35
  [narrow] 2h 35s
  [short] 2 tim, 35 s
  [long] 2 timmar, 35 sekunder
